+++ src/components/tooltip/tooltip-content.tsx
@@ -1,4 +1,5 @@
-import { Component, h, Prop } from '@stencil/core';
+import { Component, h, Host, Prop } from '@stencil/core';
+import { normalizeHotkeyString } from '../../util/hotkeys';
 
 /**
  * This component is used internally by `limel-tooltip`.
@@ -29,6 +30,12 @@ export class TooltipContent {
     @Prop({ reflect: true })
     maxlength?: number;
 
+    /**
+     * Read more in tooltip.tsx
+     */
+    @Prop({ reflect: true })
+    hotkey?: string;
+
     public render() {
         let isLabelsTextLong = false;
         if (this.helperLabel && this.maxlength) {
@@ -36,18 +43,40 @@ export class TooltipContent {
                 this.label.length + this.helperLabel.length > this.maxlength;
         }
 
-        const props: any = {};
-        if (this.maxlength) {
-            props.style = {
-                '--tooltip-max-width-of-text': `${this.maxlength}` + 'ch',
-            };
+        const style = this.maxlength
+            ? { '--tooltip-max-width-of-text': `${this.maxlength}ch` }
+            : undefined;
+
+        return (
+            <Host
+                class={{ 'has-column-layout': isLabelsTextLong }}
+                style={style}
+            >
+                <div class="label">{this.label}</div>
+                {this.renderHelperLabel()}
+                {this.renderHotkey()}
+            </Host>
+        );
+    }
+
+    private renderHelperLabel() {
+        if (!this.helperLabel) {
+            return;
         }
 
-        return [
-            <text class={{ 'has-column-layout': isLabelsTextLong }} {...props}>
-                <div class="label">{this.label}</div>
-                <div class="helper-label">{this.helperLabel}</div>
-            </text>,
-        ];
+        return <div class="helper-label">{this.helperLabel}</div>;
+    }
+
+    private renderHotkey() {
+        if (!this.hotkey) {
+            return;
+        }
+
+        const normalized = normalizeHotkeyString(this.hotkey);
+        if (!normalized) {
+            return;
+        }
+
+        return <limel-hotkey value={normalized} />;
     }
 }